fix(trusty-mpm): split oversized files to satisfy the 500-SLOC line-cap gate
CI's line-cap check flagged two files after the session-context tools PR:
catchup/mod.rs grew to 601 SLOC (new, not allowlisted) and
daemon/mcp_backend.rs grew to 600 SLOC (past its frozen 566 budget).

- Extract the structured-catchup JSON types + generate_catchup_json (and
  their tests) from catchup/mod.rs into a new catchup/json.rs sibling.
- Extract mcp_backend.rs's inline #[cfg(test)] mod tests into
  mcp_backend_tests.rs (the `_tests.rs` naming exempts it from the
  production cap, matching the doctor.rs/doctor_tests.rs precedent).
- Ratchet .line-cap-allowlist.tsv down: mcp_backend.rs dropped back under
  the 500 cap so its allowlist entry is removed.

No behavior change; `cargo test -p trusty-common --features catchup` and
`cargo test -p trusty-mpm` still pass with the same counts.
